♻️ refactor(quota): move quota page into feature module
Move the quota route implementation and shared styles into src/features/quota.

The route wrapper stays in pages while quota components reference feature-owned styles.
